feat-enhance JWT-based authentication and user registration
Added DTOs for authentication and registration requests and responses. Introduced AuthService for handling user registration and authentication logic. Refactored JWT service and filter, moved classes to appropriate packages, and updated controller dependencies. Enhanced Role and User entities for role management and user ID type. Added DemoController for secured endpoint testing.

+++ src/main/java/com/TM/taskmanager/entity/User.java
@@ -1,6 +1,7 @@
 package com.TM.taskmanager.entity;
 
 import jakarta.persistence.*;
+import jakarta.persistence.criteria.CriteriaBuilder.In;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -21,7 +22,7 @@ import java.util.List;
 public class User implements UserDetails {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     private String password;
